Fix achievements page URL
Fix an issue were some of the achievement page urls were given to the wrong game, leading to some games to have the wrong achievement stats.

This was caused by the find_all() method not finding the correct number of buttons for the buttons_links variable. Usually, games without achievements or leaderboards don't have a stats button, so the number of buttons on one game element has one button that is both a div element with attribute class="pullup_item".