Fix incorrect average game completion rate
Fix a bug where the average game completion rate was being calculated wrong. This bug was caused the SQLite database incorrectly storing some games twice with different achievement percentages.

The fix involves iterating through the outputted list which contained all the games with achievements (including the corrupted game achievements) in the SQLite database. Only the games with the correct achievement percentage were taken and duplicates where removed using a set.

This fix is not efficient as it replaces the incorrect game percentages with another copy of the correct percentage resulting in duplicates. Hence, a set had to be used. Also, this does not prevent the issue of why the data is incorrect in the first place.