Fix security code error
Attempt to fix a bug that occurs after entering the correct email security code when signing in. This error is caused by the wrong security code function when checking whether the authentication menu is still displayed.

However, when successfully logging into the steam the isdisplay() function seems to not be able to return false when looking whether the user is still on the authentication menu.

The fix instead looks at the URL to know whether the user is still stuck at the log-in page or has succesfully logged into Steam. There maybe a better solution to this bug.