Fix security code error 2
Fix the "element cannot be found" error when attempting to click the proceed button after re-entering the security code. This was caused by the proceed button not being found as it was still being loaded when trying to click it.

The fix waits up to 5 seconds for the proceed button element to become clickable before attempting to click it.