Multiple security code attempts
Update main.py to allow multiple attempts for entering the security code when the user is signing into Steam. A while statement loops around to check whether the user has successfully signed in or their entered security code is incorrect. If the code is wrong, then the loop starts over again asking for the correct code again until it's right.