Create directory and steam_web_scraper module
Create a new steam_web_scraper module that contains some of the web scraping codes. This module contains a new function that asks the user for the email codes until its correct. Also, refactor the steam log in function to use less lines of code.

Create a new directory to hold all of the Python code.