Add navigation function to steam_web_scraper module
Add a new function to steam_web_scraper module called go_to_games_page() to navigate from the steam login page to the user's game page. This function returns an output message to notify the user's that the chrome driver is in the process of navigating.

Also, refactor some comments in main.py to better explain some code.