Add example game data
Add a small example of game data to demostrate the application without using a steam profile. This feature does not scrape data from the user's profile but instead takes data from a list.

The app gives the user's two options. Option 1 is use their steam profile to scrape data or option 2 to use the example data.