Fix incorrect email code bug
Fix a bug where the app sometimes recognises the email code as incorrect when the correct code was given, causing the app to ask for the code again.

This bug is caused by slow internet connection. The app checks whether the security code worked by looking at whether the web address is still at the login page or has logged into their profile page after 3 seconds. This set amaount of time is increased to 5 seconds to allow enough time for slow browsers.