Add playtime SQLite column
Add a new column to the SQLite table that holds the number of hours that the player has played for each game. This data is represented as a string that was created by stripping the hours_played string variable in the steam_scraper.py module.

Also, add play time data for each game in the demo game_data_list in main.py.