Fix play time attribute error
Fix an attribute error that occured when scraping the play time for each game. Games that haven't been played don't have their play time specified unlike played games.

This results in the beautifulsoup4 find method to return None instead of a string. The attribute error arises when concatenated with the achievement stats string.